modules/requirements.txt
openai
python-dotenv
requests
pandas
beautifulsoup4
pytest
pillow
deep-translator
pytrends
paapi5-python-sdk


